+++ client.go
@@ -1,25 +1,95 @@
 package wishlist
 
 import (
-	"errors"
 	"fmt"
 	"io"
 	"log"
-	"net"
 	"os"
+	"os/signal"
 	"os/user"
 	"path/filepath"
+	"syscall"
 
-	"github.com/charmbracelet/keygen"
 	"github.com/gliderlabs/ssh"
 	"github.com/muesli/termenv"
 	gossh "golang.org/x/crypto/ssh"
-	"golang.org/x/crypto/ssh/agent"
-	"golang.org/x/crypto/ssh/knownhosts"
 	"golang.org/x/crypto/ssh/terminal"
+	"golang.org/x/term"
 )
 
-func connectLocal(e *Endpoint) error {
+// TODO: move this elsewhere
+
+type sshClient interface {
+	Connect(*Endpoint) error
+}
+
+type remoteClient struct {
+	session ssh.Session
+	stdin   io.Reader
+}
+
+func (c *remoteClient) Connect(e *Endpoint) error {
+	resetPty(c.session)
+
+	method, closers, err := remoteBestAuthMethod(c.session)
+	defer closers.close()
+	if err != nil {
+		return fmt.Errorf("failed to find an auth method: %w", err)
+	}
+
+	conf := &gossh.ClientConfig{
+		User:            firstNonEmpty(e.User, c.session.User()),
+		HostKeyCallback: hostKeyCallback(e, ".wishlist/known_hosts"),
+		Auth:            []gossh.AuthMethod{method},
+	}
+
+	session, cl, err := createSession(conf, e)
+	defer cl.close()
+	if err != nil {
+		return fmt.Errorf("failed to create session: %w", err)
+	}
+
+	log.Printf("%s connect to %q, %s", c.session.User(), e.Name, c.session.RemoteAddr().String())
+
+	session.Stdout = c.session
+	session.Stderr = c.session.Stderr()
+	session.Stdin = c.stdin
+
+	pty, winch, _ := c.session.Pty()
+	w := pty.Window
+	if err := session.RequestPty(pty.Term, w.Height, w.Width, nil); err != nil {
+		return fmt.Errorf("failed to request pty: %w", err)
+	}
+
+	done := make(chan bool, 1)
+	defer func() { done <- true }()
+
+	go c.notifyWindowChanges(session, done, winch)
+
+	return shellAndWait(session)
+}
+
+func (c *remoteClient) notifyWindowChanges(session *gossh.Session, done <-chan bool, winch <-chan ssh.Window) {
+	for {
+		select {
+		case <-done:
+			return
+		case w := <-winch:
+			if w.Height == 0 && w.Width == 0 {
+				// this only happens if the session is already dead, make sure there are no leftovers
+				return
+			}
+			if err := session.WindowChange(w.Height, w.Width); err != nil {
+				log.Println("failed to notify window change:", err)
+				return
+			}
+		}
+	}
+}
+
+type localClient struct{}
+
+func (c *localClient) Connect(e *Endpoint) error {
 	resetPty(os.Stdout)
 	defer resetPty(os.Stdout)
 
@@ -43,119 +113,65 @@ func connectLocal(e *Endpoint) error {
 		HostKeyCallback: hostKeyCallback(e, filepath.Join(home, ".ssh/known_hosts")),
 	}
 
-	conn, err := gossh.Dial("tcp", e.Address, conf)
+	session, cls, err := createSession(conf, e)
+	defer cls.close()
 	if err != nil {
-		return fmt.Errorf("connection failed: %w", err)
+		return fmt.Errorf("failed to create sessio: %w", err)
 	}
-	defer func() {
-		if err := conn.Close(); err != nil {
-			log.Println("failed to close conn:", err)
-		}
-	}()
-
-	session, err := conn.NewSession()
-	if err != nil {
-		return fmt.Errorf("failed to open session: %w", err)
-	}
-
-	defer func() {
-		if err := session.Close(); err != nil && err != io.EOF {
-			log.Println("failed to close session:", err)
-		}
-	}()
 
 	session.Stdout = os.Stdout
 	session.Stderr = os.Stderr
 	session.Stdin = os.Stdin
 
-	w, h, err := terminal.GetSize(0)
+	w, h, err := terminal.GetSize(int(os.Stdout.Fd()))
 	if err != nil {
 		return fmt.Errorf("failed to get term size: %w", err)
 	}
 
-	if err := session.RequestPty("xterm-256", h, w, nil); err != nil {
+	if err := session.RequestPty(os.Getenv("$TERM"), h, w, nil); err != nil {
 		return fmt.Errorf("failed to request a pty: %w", err)
 	}
-	// TODO: handle resizes... somehow
-	if err := session.Shell(); err != nil {
-		return fmt.Errorf("failed to start shell: %w", err)
-	}
 
-	if err := session.Wait(); err != nil {
-		return fmt.Errorf("session failed: %w", err)
-	}
-	return nil
-
-}
-
-func resetPty(w io.Writer) {
-	fmt.Fprint(w, termenv.CSI+termenv.ExitAltScreenSeq)
-	fmt.Fprint(w, termenv.CSI+termenv.ResetSeq+"m")
-	fmt.Fprintf(w, termenv.CSI+termenv.EraseDisplaySeq, 2) // nolint:gomnd
-}
-
-func mustConnect(s ssh.Session, e *Endpoint, stdin io.Reader) {
-	if err := connect(s, e, stdin); err != nil {
-		fmt.Fprintf(s, "wishlist: %s\n\r", err.Error())
-		_ = s.Exit(1)
-		return // unreachable
-	}
-	fmt.Fprintf(s, "wishlist: closed connection to %q (%s)\n\r", e.Name, e.Address)
-	_ = s.Exit(0)
-}
-
-func connect(prev ssh.Session, e *Endpoint, stdin io.Reader) error {
-	resetPty(prev)
-
-	method, closers, err := authMethod(prev)
-	defer closers.close()
-	if err != nil {
-		return fmt.Errorf("failed to find an auth method: %w", err)
-	}
-
-	conf := &gossh.ClientConfig{
-		User:            firstNonEmpty(e.User, prev.User()),
-		HostKeyCallback: hostKeyCallback(e, ".wishlist/known_hosts"),
-		Auth:            []gossh.AuthMethod{method},
-	}
-
-	conn, err := gossh.Dial("tcp", e.Address, conf)
-	if err != nil {
-		return fmt.Errorf("connection failed: %w", err)
-	}
+	sig := make(chan os.Signal, 1)
+	signal.Notify(sig, syscall.SIGWINCH)
 	defer func() {
-		if err := conn.Close(); err != nil {
-			log.Println("failed to close conn:", err)
+		signal.Stop(sig)
+	}()
+
+	go func() {
+		for {
+			select {
+			case <-sig:
+				w, h, err := term.GetSize(int(os.Stdout.Fd()))
+				if err != nil {
+					log.Println(err)
+				}
+				session.WindowChange(h, w)
+			}
 		}
 	}()
+
+	return shellAndWait(session)
+}
+
+func createSession(conf *gossh.ClientConfig, e *Endpoint) (*gossh.Session, closers, error) {
+	var cl closers
+	conn, err := gossh.Dial("tcp", e.Address, conf)
+	if err != nil {
+		return nil, cl, fmt.Errorf("connection failed: %w", err)
+	}
+
+	cl = append(cl, conn.Close)
 
 	session, err := conn.NewSession()
 	if err != nil {
-		return fmt.Errorf("failed to open session: %w", err)
+		return nil, cl, fmt.Errorf("failed to open session: %w", err)
 	}
+	cl = append(cl, session.Close)
+	return session, cl, nil
+}
 
-	log.Printf("%s connect to %q, %s -> %s", prev.User(), e.Name, prev.RemoteAddr().String(), conn.RemoteAddr().String())
-	defer func() {
-		if err := session.Close(); err != nil && err != io.EOF {
-			log.Println("failed to close session:", err)
-		}
-	}()
-
-	session.Stdout = prev
-	session.Stderr = prev.Stderr()
-	session.Stdin = stdin
-
-	pty, winch, _ := prev.Pty()
-	w := pty.Window
-	if err := session.RequestPty(pty.Term, w.Height, w.Width, nil); err != nil {
-		return fmt.Errorf("failed to request pty: %w", err)
-	}
-
-	done := make(chan bool, 1)
-	defer func() { done <- true }()
-
-	go notifyWindowChanges(session, done, winch)
-
+func shellAndWait(session *gossh.Session) error {
 	if err := session.Shell(); err != nil {
 		return fmt.Errorf("failed to start shell: %w", err)
 	}
@@ -164,24 +180,6 @@ func connect(prev ssh.Session, e *Endpoint, stdin io.Reader) error {
 		return fmt.Errorf("session failed: %w", err)
 	}
 	return nil
-}
-
-func notifyWindowChanges(session *gossh.Session, done <-chan bool, winch <-chan ssh.Window) {
-	for {
-		select {
-		case <-done:
-			return
-		case w := <-winch:
-			if w.Height == 0 && w.Width == 0 {
-				// this only happens if the session is already dead, make sure there are no leftovers
-				return
-			}
-			if err := session.WindowChange(w.Height, w.Width); err != nil {
-				log.Println("failed to notify window change:", err)
-				return
-			}
-		}
-	}
 }
 
 type closers []func() error
@@ -194,89 +192,6 @@ func (c closers) close() {
 	}
 }
 
-// authMethod returns an auth method.
-//
-// it first tries to use ssh-agent, if that's not available, it creates and uses a new key pair.
-func authMethod(s ssh.Session) (gossh.AuthMethod, closers, error) {
-	method, closers, err := tryAuthAgent(s)
-	if err != nil {
-		return method, closers, err
-	}
-	if method != nil {
-		return method, closers, nil
-	}
-
-	method, err = tryNewKey()
-	return method, closers, err
-}
-
-// tryAuthAgent will try to use an ssh-agent to authenticate.
-func tryAuthAgent(s ssh.Session) (gossh.AuthMethod, closers, error) {
-	_, _ = s.SendRequest("auth-agent-req@openssh.com", true, nil)
-
-	if ssh.AgentRequested(s) {
-		l, err := ssh.NewAgentListener()
-		if err != nil {
-			return nil, nil, err // nolint:wrapcheck
-		}
-		go ssh.ForwardAgentConnections(l, s)
-
-		conn, err := net.Dial(l.Addr().Network(), l.Addr().String())
-		if err != nil {
-			return nil, closers{l.Close}, err // nolint:wrapcheck
-		}
-
-		return gossh.PublicKeysCallback(agent.NewClient(conn).Signers),
-			closers{l.Close, conn.Close},
-			nil
-	}
-
-	fmt.Fprintf(s.Stderr(), "wishlist: ssh agent not available\n\r")
-	return nil, nil, nil
-}
-
-func tryUserKeys(home string) ([]gossh.AuthMethod, error) {
-	var methods []gossh.AuthMethod
-	for _, name := range []string{
-		"id_rsa",
-		"id_ed25519",
-	} {
-		path := filepath.Join(home, ".ssh", name)
-		bts, err := os.ReadFile(path)
-		if err != nil {
-			continue
-		}
-		signer, err := gossh.ParsePrivateKey(bts)
-		if err != nil {
-			return methods, err
-		}
-		log.Printf("using %q", path)
-		methods = append(methods, gossh.PublicKeys(signer))
-	}
-
-	return methods, nil
-}
-
-// tryNewKey will create a .wishlist/client_ed25519 keypair if one does not exist.
-// It will return an auth method that uses the keypair if it exist or is successfully created.
-func tryNewKey() (gossh.AuthMethod, error) {
-	key, err := keygen.New(".wishlist", "client", nil, keygen.Ed25519)
-	if err != nil {
-		return nil, err // nolint:wrapcheck
-	}
-
-	signer, err := gossh.ParsePrivateKey(key.PrivateKeyPEM)
-	if err != nil {
-		return nil, err // nolint:wrapcheck
-	}
-
-	if key.IsKeyPairExists() {
-		return gossh.PublicKeys(signer), nil
-	}
-
-	return gossh.PublicKeys(signer), key.WriteKeys()
-}
-
 func firstNonEmpty(ss ...string) string {
 	for _, s := range ss {
 		if s != "" {
@@ -286,35 +201,8 @@ func firstNonEmpty(ss ...string) string {
 	return ""
 }
 
-// hostKeyCallback returns a callback that will be used to verify the host key.
-//
-// it creates a file in the given path, and uses that to verify hosts and keys.
-// if the host does not exist there, it adds it so its available next time, as plain old `ssh` does.
-func hostKeyCallback(e *Endpoint, path string) gossh.HostKeyCallback {
-	return func(hostname string, remote net.Addr, key gossh.PublicKey) error {
-		kh, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0o600) // nolint:gomnd
-		if err != nil {
-			return fmt.Errorf("failed to open known_hosts: %w", err)
-		}
-		defer func() { _ = kh.Close() }()
-
-		callback, err := knownhosts.New(kh.Name())
-		if err != nil {
-			return fmt.Errorf("failed to check known_hosts: %w", err)
-		}
-
-		if err := callback(hostname, remote, key); err != nil {
-			var kerr *knownhosts.KeyError
-			if errors.As(err, &kerr) {
-				if len(kerr.Want) > 0 {
-					return fmt.Errorf("possible man-in-the-middle attack: %w", err)
-				}
-				// if want is empty, it means the host was not in the known_hosts file, so lets add it there.
-				fmt.Fprintln(kh, knownhosts.Line([]string{e.Address}, key))
-				return nil
-			}
-			return fmt.Errorf("failed to check known_hosts: %w", err)
-		}
-		return nil
-	}
+func resetPty(w io.Writer) {
+	fmt.Fprint(w, termenv.CSI+termenv.ExitAltScreenSeq)
+	fmt.Fprint(w, termenv.CSI+termenv.ResetSeq+"m")
+	fmt.Fprintf(w, termenv.CSI+termenv.EraseDisplaySeq, 2) // nolint:gomnd
 }
