+++ go.mod
@@ -14,5 +14,6 @@ require (
 	github.com/muesli/termenv v0.9.0
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/crypto v0.0.0-20220112180741-5e0467b6c7ce
+	golang.org/x/term v0.0.0-20210422114643-f5beecf764ed
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
