+++ middleware.go
@@ -103,3 +103,17 @@ func listenAppEvents(s ssh.Session, p *tea.Program, donech <-chan bool, errch <-
 		}
 	}
 }
+
+func mustConnect(s ssh.Session, e *Endpoint, stdin io.Reader) {
+	client := &remoteClient{
+		session: s,
+		stdin:   stdin,
+	}
+	if err := client.Connect(e); err != nil {
+		fmt.Fprintf(s, "wishlist: %s\n\r", err.Error())
+		_ = s.Exit(1)
+		return // unreachable
+	}
+	fmt.Fprintf(s, "wishlist: closed connection to %q (%s)\n\r", e.Name, e.Address)
+	_ = s.Exit(0)
+}
